correct lstmap

+++ ft_strcat.c
@@ -6,7 +6,7 @@
 /*   By: wvenita <wvenita@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/05 21:53:27 by wvenita           #+#    #+#             */
-/*   Updated: 2019/04/15 23:38:43 by wvenita          ###   ########.fr       */
+/*   Updated: 2019/04/17 23:07:56 by wvenita          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strcat(char *dest, const char *src)
 {
-	int		i;
-	int		n;
+	int			i;
+	size_t		n;
 
 	i = -1;
 	n = ft_strlen(dest);
